Fix database configurations
