+++ search.service.search.duftstoffident/src/main/java/de/hswt/fi/search/service/search/duftstoffident/DuftstoffidentCompoundService.java
@@ -1,10 +1,10 @@
-package de.hswt.fi.search.service.search.plantident;
+package de.hswt.fi.search.service.search.duftstoffident;
 
 import de.hswt.fi.common.spring.Profiles;
 import de.hswt.fi.search.service.search.api.AbstractCompoundService;
 import de.hswt.fi.search.service.search.api.repositories.EntryRepository;
-import de.hswt.fi.search.service.search.plantident.config.DuftStoffIdentDatabaseConfiguration;
-import de.hswt.fi.search.service.search.plantident.repositories.DuftstoffidentEntryRepository;
+import de.hswt.fi.search.service.search.duftstoffident.config.DuftStoffIdentDatabaseConfiguration;
+import de.hswt.fi.search.service.search.duftstoffident.repositories.DuftstoffidentEntryRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
