+++ search.service.search.duftstoffident/src/main/java/de/hswt/fi/search/service/search/duftstoffident/config/DuftStoffIdentDatabaseConfiguration.java
@@ -1,4 +1,4 @@
-package de.hswt.fi.search.service.search.plantident.config;
+package de.hswt.fi.search.service.search.duftstoffident.config;
 
 import de.hswt.fi.common.spring.Profiles;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -19,26 +19,28 @@ import javax.persistence.EntityManagerFactory;
 import javax.persistence.PersistenceUnit;
 import javax.sql.DataSource;
 
-@Profile({Profiles.LC, Profiles.DEVELOPMENT_LC})
+import static de.hswt.fi.search.service.search.duftstoffident.config.DuftStoffIdentDatabaseConfiguration.*;
+
+@Profile({Profiles.GC, Profiles.DEVELOPMENT_GC})
 @Configuration
 @EnableTransactionManagement
 @EnableJpaRepositories(
-		basePackages = StoffIdentDatabaseConfiguration.BASE_PACKAGE,
-		entityManagerFactoryRef = StoffIdentDatabaseConfiguration.ENTITY_MANAGER,
-		transactionManagerRef = StoffIdentDatabaseConfiguration.TRANSACTION_MANAGER
+		basePackages = BASE_PACKAGE,
+		entityManagerFactoryRef = ENTITY_MANAGER,
+		transactionManagerRef = TRANSACTION_MANAGER
 )
-public class StoffIdentDatabaseConfiguration {
+public class DuftStoffIdentDatabaseConfiguration {
 
-	public static final String ID_PREFIX = "CP";
-	public static final String DATABASE_NAME = "STOFF-IDENT";
-	public static final String ENTITY_MANAGER = "stoffidentEntityManager";
-	public static final String TRANSACTION_MANAGER = "stoffidentTransactionManager";
-	static final String BASE_PACKAGE = "de.hswt.fi.search.service.search.stoffident.repositories";
-	private static final String CONFIGURATION_PREFIX = "spring.datasource.stoffident";
-	private static final String PERSISTENCE_UNIT = "stoffidentPersistenceUnit";
+	public static final String ID_PREFIX = "DI";
+	public static final String DATABASE_NAME = "DUFTSTOFF-IDENT";
+	public static final String ENTITY_MANAGER = "duftstoffidentEntityManager";
+	public static final String TRANSACTION_MANAGER = "duftstoffidentTransactionManager";
+	static final String BASE_PACKAGE = "de.hswt.fi.search.service.search.duftstoffident.repositories";
+	private static final String CONFIGURATION_PREFIX = "spring.datasource.duftstoffident";
+	private static final String PERSISTENCE_UNIT = "duftstoffidentPersistenceUnit";
 	private static final String ENTITY_PACKAGES = "de.hswt.fi.search.service.mass.search.model";
-	private static final String DATA_SOURCE = "stoffidentDataSource";
-	private static final String DATA_SOURCE_PROPERTIES = "stoffidentDataSourceProperties";
+	private static final String DATA_SOURCE = "duftstoffidentDataSource";
+	private static final String DATA_SOURCE_PROPERTIES = "duftstoffidentDataSourceProperties";
 
 	@Bean(name = DATA_SOURCE_PROPERTIES)
 	@ConfigurationProperties(CONFIGURATION_PREFIX)
@@ -47,13 +49,13 @@ public class StoffIdentDatabaseConfiguration {
 	}
 
 	@Bean(name = DATA_SOURCE)
-	public DataSource stoffidentDataSource() {
+	public DataSource duftstoffidentDataSource() {
 		return dataSourceProperties().initializeDataSourceBuilder().build();
 	}
 
 	@Bean(name = ENTITY_MANAGER)
 	@PersistenceUnit(name = PERSISTENCE_UNIT)
-	public LocalContainerEntityManagerFactoryBean stoffidentEntityManager(@Qualifier(DATA_SOURCE) DataSource dataSource,
+	public LocalContainerEntityManagerFactoryBean duftstoffidentEntityManager(@Qualifier(DATA_SOURCE) DataSource dataSource,
 																			  // Fix for generate-ddl property not picked up by spring with multiple data sources
 																			  @Value("${" + CONFIGURATION_PREFIX + ".generate-ddl:false}") boolean generateDdl) {
 		LocalContainerEntityManagerFactoryBean em = new LocalContainerEntityManagerFactoryBean();
@@ -70,7 +72,7 @@ public class StoffIdentDatabaseConfiguration {
 	}
 
 	@Bean(name = TRANSACTION_MANAGER)
-	public PlatformTransactionManager stoffidentTransactionManager(@Qualifier(ENTITY_MANAGER) EntityManagerFactory entityManagerFactory) {
+	public PlatformTransactionManager duftstoffidentransactionManager(@Qualifier(ENTITY_MANAGER) EntityManagerFactory entityManagerFactory) {
 		return new JpaTransactionManager(entityManagerFactory);
 	}
 }
