+++ search.service.search.duftstoffident/src/main/java/de/hswt/fi/search/service/search/duftstoffident/repositories/DuftstoffidentEntryRepository.java
@@ -1,4 +1,4 @@
-package de.hswt.fi.search.service.search.plantident.repositories;
+package de.hswt.fi.search.service.search.duftstoffident.repositories;
 
 import de.hswt.fi.search.service.search.api.repositories.EntryRepository;
